Greengrass V2 API polling example
This example shows how to use Greengrass V2 to poll a local HTTP/HTTPS API and optionally filter, process, and store the data in S3.

The GGUtils component rolls up a lot of boilerplate code into a utility class so working with streams requires less code. It also implements context managers for the Greengrass stream manager and the Greengrass IPC class so they can clean up resources automatically with Python's 'with' statement.